Refactored code for Homepage
