+++ src/components/carousel/hero-carousel-left.tsx
