+++ src/components/carousel/hero-carousel-middle.tsx
