+++ src/index.css
@@ -130,3 +130,7 @@
 .button3 {
   @apply mt-4 bg-white text-black text-xl px-6 py-6 rounded-sm hover:text-white hover:bg-transparent border hover:border-white active:translate-y-0.5 transition-all;
 }
+
+.header1 {
+  @apply text-3xl font-medium mb-6;
+}
