+++ src/pages/home/homepage.tsx
@@ -1,17 +1,14 @@
-import HeroSection from "../../components/herosection/herosection";
-import MostPopularBooks from "../../components/herosection/most-popular-books";
+import FeaturedCategories from "../../components/homepage/featured-categories";
+import HeroSection from "../../components/homepage/herosection";
+import MostPopularBooks from "../../components/homepage/most-popular-books";
 
 import "../../index.css";
 const Homepage = () => {
   return (
-    <div className="gap-8">
+    <div className="px-12 space-y-20">
       <HeroSection />
-      <div className="p-12 space-y-6">
-        <h2 className="text-3xl font-medium">
-          Most Popular For Food and Drinks books
-        </h2>
-        <MostPopularBooks />
-      </div>
+      <MostPopularBooks />
+      <FeaturedCategories />
     </div>
   );
 };
